+++ protocols/http_probe.go
@@ -13,7 +13,7 @@
 //
 //    with status 301
 //
-// You can allow multiple statuses by joining them with a comma:
+// You can also allow multiple statuses by joining them with a comma:
 //
 //    with status 200,429
 //
@@ -304,6 +304,10 @@ func (s *HTTPTest) RunTest(tst test.Test, target string, opts test.Options) erro
 	//
 	address := target
 	target = tst.Target
+
+	// If we're running a period test, check if we can replace any related vars
+	target = strings.Replace(target, "__pt-index__", strconv.Itoa(opts.PeriodTestIndex), -1)
+	target = strings.Replace(target, "__pt-time-ms__", strconv.FormatInt(opts.PeriodTestStartTime, 10), -1)
 
 	//
 	// Setup a dialer which will be dual-stack
@@ -680,12 +684,12 @@ func (s *HTTPTest) RunTest(tst test.Test, target string, opts test.Options) erro
 		//
 		// Check the expiration
 		//
-		hours, errExpire := s.SSLExpiration(tst.Target, opts.Verbose)
+		hours, cn, errExpire := s.SSLExpiration(tst.Target, opts.Verbose)
 		if errExpire == nil {
 			// Is the age too short?
 			if int64(hours) < int64(period) {
 
-				return fmt.Errorf("SSL certificate will expire in %d hours (%d days)", hours, int(hours/24))
+				return fmt.Errorf("SSL certificate '%s' will expire in %d hours (%d days)", cn, hours, int(hours/24))
 			}
 		}
 
@@ -699,18 +703,21 @@ func (s *HTTPTest) RunTest(tst test.Test, target string, opts test.Options) erro
 
 // SSLExpiration returns the number of hours remaining for a given
 // SSL certificate chain.
-func (s *HTTPTest) SSLExpiration(host string, verbose bool) (int64, error) {
+func (s *HTTPTest) SSLExpiration(host string, verbose bool) (int64, string, error) {
 
 	// Expiry time, in hours
 	var hours int64
 	hours = -1
+
+	// The common-name of the certificate involved.
+	cn := ""
 
 	//
 	// If the string matches https://, then strip it off
 	//
 	re, err := regexp.Compile(`^https:\/\/([^\/]+)`)
 	if err != nil {
-		return 0, err
+		return 0, "", err
 	}
 	res := re.FindAllStringSubmatch(host, -1)
 	for _, v := range res {
@@ -734,7 +741,7 @@ func (s *HTTPTest) SSLExpiration(host string, verbose bool) (int64, error) {
 
 	conn, err := tls.Dial("tcp", host, nil)
 	if err != nil {
-		return 0, err
+		return 0, "", err
 	}
 	defer conn.Close()
 
@@ -752,18 +759,24 @@ func (s *HTTPTest) SSLExpiration(host string, verbose bool) (int64, error) {
 			// If we've not checked anything this is the benchmark
 			if hours == -1 {
 				hours = expiresIn
+				cn = cert.Subject.CommonName
 			} else {
 				// Otherwise replace our result if the
 				// certificate is going to expire more
 				// recently than the current "winner".
 				if expiresIn < hours {
 					hours = expiresIn
+					cn = cert.Subject.CommonName
 				}
 			}
 		}
 	}
 
-	return hours, nil
+	return hours, cn, nil
+}
+
+func (s *HTTPTest) GetUniqueHashForTest(tst test.Test, opts test.Options) *string {
+	return nil
 }
 
 // init is used to dynamically register our protocol-tester.